requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
sendgrid==6.11.0
twilio==9.0.0
